Support for reading form data in ajp13 and some bug fixes
git-svn-id: https://svn.apache.org/repos/asf/tomcat@281939 13f79535-47bb-0310-9956-ffa450edef68

+++ container/tc3.x/trunk/src/native/jk/jk_ajp13.c
@@ -353,7 +353,7 @@ int ajp13_marshal_into_msgb(jk_msg_buf_t *msg,
        0 != jk_b_append_string(msg, s->remote_addr)     ||
        0 != jk_b_append_string(msg, s->remote_host)     ||
        0 != jk_b_append_string(msg, s->server_name)     ||
-       0 != jk_b_append_int(msg, s->server_port)        ||
+       0 != jk_b_append_int(msg, (unsigned short)s->server_port) ||
        0 != jk_b_append_byte(msg, (unsigned char)(s->is_ssl)) ||
        0 != jk_b_append_int(msg, (unsigned short)(s->num_headers))) {
         return JK_FALSE;
